requirements.txt
click 
pyzmq 
sentence-transformers
fastapi